+++ nautobot_circuit_maintenance/migrations/0003_improve_rawnotification.py
@@ -4,6 +4,17 @@ from django.db import migrations, models
 import django.db.models.deletion
 
 
+def migrate_source(apps, schema_editor):
+    """Migrate from old text Source to new reference to Notification Source."""
+
+    RawNotificationModel = apps.get_model("nautobot_circuit_maintenance", "RawNotification")
+    NotificationSourceModel = apps.get_model("nautobot_circuit_maintenance", "NotificationSource")
+
+    for raw_notification in RawNotificationModel.objects.all():
+        raw_notification.source = NotificationSourceModel.objects.get(name=raw_notification.source_old)
+        raw_notification.save()
+
+
 class Migration(migrations.Migration):
 
     dependencies = [
@@ -16,7 +27,12 @@ class Migration(migrations.Migration):
             name="sender",
             field=models.CharField(blank=True, default="", max_length=200, null=True),
         ),
-        migrations.AlterField(
+        migrations.RenameField(
+            model_name="rawnotification",
+            old_name="source",
+            new_name="source_old",
+        ),
+        migrations.AddField(
             model_name="rawnotification",
             name="source",
             field=models.ForeignKey(
@@ -24,5 +40,10 @@ class Migration(migrations.Migration):
                 on_delete=django.db.models.deletion.SET_NULL,
                 to="nautobot_circuit_maintenance.notificationsource",
             ),
+        ),
+        migrations.RunPython(migrate_source),
+        migrations.RemoveField(
+            model_name="rawnotification",
+            name="source_old",
         ),
     ]